ReStruct AIManager, AIFreeBattle mode has created. And fix the algorithm of judging move target.
