+++ Assets/Scripts/Core/AI/AIFreeBattle.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 87807ac09cd9a7c46851cfb246db2376
+timeCreated: 1524663039
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
