--- Assets/Scripts/Core/AI/Drama.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public abstract class Drama : MonoBehaviour {
-
-    public abstract IEnumerator Play();
-
-}
