+++ Assets/Scripts/Core/AI/Drama.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: c5466db4a3c31c345bdec1e48ffa6057
+folderAsset: yes
+timeCreated: 1524674637
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
