+++ Assets/Scripts/Core/AI/Drama/Drama.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 8727530c18b814e4f9d889a41435e105
+timeCreated: 1524574116
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
