+++ Assets/Scripts/Core/AI/Drama/DramaBox.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: e155824551436e449a04260972ce6a35
+timeCreated: 1524674677
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
