+++ Assets/Scripts/Core/AI/Drama/DramaNinjaCombo.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DramaNinjaComb : Drama {
+    public override IEnumerator Play()
+    {
+        throw new System.NotImplementedException();
+    }
+}
