+++ Assets/Scripts/Core/AI/Drama/DramaNinjaCombo.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 3b7ac2c5270582643a9bebd521c56ed9
+timeCreated: 1524674775
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
