+++ Assets/Scripts/Core/Player/HumanPlayer.cs
@@ -1,4 +1,8 @@
 using UnityEngine;
+
+/// <summary>
+/// Human Round
+/// </summary>
 class HumanPlayer : Player
 {
     public override void Play(RoundManager roundManager)
